seeder

+++ database/migrations/2021_02_18_155219_add_discount_price_to_products_table.php
@@ -14,7 +14,7 @@ class AddDiscountPriceToProductsTable extends Migration
     public function up()
     {
         Schema::table('products', function (Blueprint $table) {
-            $table->integer('discount_price')->after('price')->nullable();
+            $table->double('discount_price')->after('price')->nullable();
         });
     }
